Change int16_t MbType and uint8_t SubMbType to uint32_t to support B-Slice.

+++ codec/decoder/core/inc/dec_frame.h
@@ -63,7 +63,7 @@ struct TagDqLayer {
 
   PBitStringAux                 pBitStringAux;  // pointer to SBitStringAux
   PFmo                          pFmo;           // Current fmo context pointer used
-  int16_t* pMbType;
+  uint32_t* pMbType;
   int32_t* pSliceIdc;                           // using int32_t for slice_idc
   int16_t (*pMv[LIST_A])[MB_BLOCK4x4_NUM][MV_A];
   int16_t (*pMvd[LIST_A])[MB_BLOCK4x4_NUM][MV_A];
@@ -86,7 +86,7 @@ struct TagDqLayer {
   uint8_t  *pIntraNxNAvailFlag;
   int8_t*  pChromaPredMode;
   //uint8_t (*motion_pred_flag[LIST_A])[MB_PARTITION_SIZE]; // 8x8
-  int8_t (*pSubMbType)[MB_SUB_PARTITION_SIZE];
+  uint32_t (*pSubMbType)[MB_SUB_PARTITION_SIZE];
   int32_t iLumaStride;
   int32_t iChromaStride;
   uint8_t* pPred[3];
@@ -130,7 +130,7 @@ typedef struct TagGpuAvcLayer {
   SLayerInfo                sLayerInfo;
   PBitStringAux             pBitStringAux;  // pointer to SBitStringAux
 
-  int16_t*                  pMbType;
+	uint32_t*                  pMbType;
   int32_t*                  pSliceIdc;      // using int32_t for slice_idc
   int8_t*                   pLumaQp;
   int8_t*                   pCbp;
