+++ codec/decoder/core/src/parse_mb_syn_cabac.cpp
@@ -419,7 +419,7 @@ int32_t ParseIntraPredModeChromaCabac (PWelsDecoderContext pCtx, uint8_t uiNeigh
   uint32_t uiCode;
   int32_t iIdxA, iIdxB, iCtxInc;
   int8_t* pChromaPredMode = pCtx->pCurDqLayer->pChromaPredMode;
-  int16_t* pMbType = pCtx->pCurDqLayer->pMbType;
+  uint32_t* pMbType = pCtx->pCurDqLayer->pMbType;
   int32_t iLeftAvail     = uiNeighAvail & 0x04;
   int32_t iTopAvail      = uiNeighAvail & 0x01;
 
@@ -991,7 +991,7 @@ int32_t ParseInterBMotionInfoCabac(PWelsDecoderContext pCtx, PWelsNeighAvail pNe
 			uint8_t iScan4Idx, iCacheIdx;
 			iCacheIdx = g_kuiCache30ScanIdx[i << 2];
 
-			int32_t subMbType = pCurDqLayer->pSubMbType[iMbXy][i];
+			uint32_t subMbType = pCurDqLayer->pSubMbType[iMbXy][i];
 			int32_t list[2] = { -1 };
 			if (pCurDqLayer->pSubMbType[iMbXy][i] == MB_TYPE_DIRECT) {
 				//Add processing for DIRECT 
@@ -1243,7 +1243,7 @@ int32_t ParseCbfInfoCabac (PWelsNeighAvail pNeighAvail, uint8_t* pNzcCache, int3
   int32_t iTopBlkXy = iCurrBlkXy - pCtx->pCurDqLayer->iMbWidth; //default value: MB neighboring
   int32_t iLeftBlkXy = iCurrBlkXy - 1; //default value: MB neighboring
   uint16_t* pCbfDc = pCtx->pCurDqLayer->pCbfDc;
-  int16_t* pMbType = pCtx->pCurDqLayer->pMbType;
+  uint32_t* pMbType = pCtx->pCurDqLayer->pMbType;
   int32_t iCtxInc;
   uiCbfBit = 0;
   nA = nB = (int8_t)!!IS_INTRA (pMbType[iCurrBlkXy]);
